Refactoring. From unsigned long to size_t

+++ philo.h
@@ -6,7 +6,7 @@
 /*   By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/11 14:28:29 by irychkov          #+#    #+#             */
-/*   Updated: 2024/11/11 23:09:45 by irychkov         ###   ########.fr       */
+/*   Updated: 2024/11/12 13:54:22 by irychkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ typedef struct s_program_data
 	const int		time_to_eat;
 	const int		time_to_sleep;
 	const int		number_of_times_each_philosopher_must_eat;
-	int				start_time;
+	size_t			start_time;
 	pthread_mutex_t	mutex_print;
 	pthread_mutex_t	*forks;
 }	t_program_data;
@@ -35,7 +35,7 @@ typedef struct s_program_data
 typedef struct s_single_philo {
 	int					id;
 	int					times_eaten;
-	unsigned long int	last_meal_time;
+	size_t				last_meal_time;
 	pthread_t			thread_id;
 	pthread_mutex_t		*left_fork;
 	pthread_mutex_t		*right_fork;
@@ -48,6 +48,6 @@ t_program_data		*init_data(int ac, char *av[]);
 t_single_philo		*init_single_philos(t_program_data *data);
 void				destroy_mutexes(t_program_data *data);
 void				free_all(t_program_data *data, t_single_philo *philos);
-unsigned long int	get_current_time(void);
+size_t				get_current_time(void);
 
 #endif
